backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.4
pydantic-settings==2.7.0
python-dotenv==1.0.1
httpx==0.28.1
requests==2.32.3
pandas==2.2.3
numpy==2.2.1
supabase==2.11.0
kiteconnect==5.0.1
python-jose[cryptography]==3.3.0
cryptography==44.0.0
